requirements.txt
adal==1.2.7
aenum==3.1.8
anyio==3.5.0
appdirs==1.4.4
APScheduler==3.8.1
asgiref==3.5.0
azure-core==1.22.1
azure-storage-blob==12.9.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
case-conversion==2.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
clipboard==0.0.4
colorama==0.4.4
colour==0.1.5
cryptography==35.0.0
cursor==1.3.4
decorator==4.4.2
dill==0.3.4
fastapi==0.73.0
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
h11==0.13.0
html-text==0.5.2
idna==3.3
imageio==2.15.0
imageio-ffmpeg==0.4.5
infi.systray==0.1.12
inflect==5.4.0
isodate==0.6.1
lxml==4.7.1
maybe-else==0.2.1
mbstrdecoder==1.1.0
moviepy==1.0.3
msal==1.16.0
msoffcrypto-tool==5.0.0
msrest==0.6.21
numpy==1.22.2
O365==2.0.18.1
oauthlib==3.2.0
office365==0.3.15
office365-rest-client==2.2.0
Office365-REST-Python-Client==2.3.8
olefile==0.46
pandas==1.4.1
parsedatetime==2.6
pathmagic==0.3.14
Pillow==9.0.1
prettierfier==1.0.3
proglog==0.1.9
pycparser==2.20
pydantic==1.9.0
pydub==0.25.1
pyinstrument==4.1.1
pyiotools==0.3.18
PyJWT==2.3.0
pymiscutils==0.3.14
PyPDF2==1.26.0
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pysubtypes==0.3.18
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
readchar==3.0.5
regex==2022.1.18
requests==2.26.0
requests-oauthlib==1.3.1
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
stringcase==1.2.0
tabulate==0.8.9
tqdm==4.62.3
typepy==1.3.0
typing-extensions==4.1.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
uvicorn==0.17.4
XlsxWriter==3.0.2
